+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This project intends to inhere to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html), but has not yet reached 1.0 so
 all APIs might be changed.
 
+## Unreleased - xxxx-xx-xx
+
+### Breaking Changes
+
+- `Error::Close` now has a code as well as a reason.
+
 ## v0.8.0-alpha.1 - 2024-01-19
 
 ### Breaking Changes